% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountShadingEvents.R
\name{CountShadingEvents}
\alias{CountShadingEvents}
\title{Count the number of shading events between sunrise and sunset.}
\usage{
CountShadingEvents(lightdata = lightdata, threshold = 1, MCMCresult = NA,
  plot = TRUE)
}
\arguments{
\item{lightdata}{an object with Date and Light.}

\item{\code{threshold}}{threshold value}

\item{\code{MCMCresult}}{either MCMC result from Estelle model or slices object}

\item{\code{plot}}{logical plot results}
}
\description{
Count the number of shading events between sunrise and sunset.
}

