% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigSchedule.R
\name{MigSchedule}
\alias{MigSchedule}
\title{Determine migratory schedule from archival light-level geolocator data.}
\usage{
MigSchedule(MCMC, prob = 0.95, mig.quantile = 0.95,
  stationary.periods = NULL, stationary.duration = 2,
  collapseSites = FALSE, rm.lat.equinox = FALSE, days.omit = 5,
  progress = TRUE, plot = TRUE, plot.legend = TRUE)
}
\arguments{
\item{MCMC}{a slices object.}

\item{\code{prob}}{probability estimate to return for daily positions}

\item{\code{mig.quantile}}{probability for determining movement periods}

\item{\code{stationary.duration}}{minimum number of stationary days to consider stationary periods, default == 2}

\item{\code{rm.lat.equinox}}{logical remove dates from consideration around equinox}

\item{\code{days.omit}}{integer how many days on either side to remove if \code{rm.lat.equniox == TRUE}}

\item{\code{collapseSites}}{logical merge stationary sites if the previous median location falls within subsequent stationary site}

\item{\code{progress}}{logical show progress bar}

\item{\code{plot}}{logical plot results upon completion}

\item{\code{plot.legend}}{logical plot legend with dates on figure}
}
\value{
\code{list} DailyPositions - geographically weighted mean and associated errors
                    Schedule - arrival and departure dates of stops
                    movements \code{rasterStack} of Schedule.
}
\description{
Determine migratory schedule from archival light-level geolocator data.
}
