% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigSchedule.R
\name{MigSchedule}
\alias{MigSchedule}
\title{Determine migratory schedule from archival light-level geolocator data.}
\usage{
MigSchedule(MCMC, prob = 0.95, known.breed = NULL, known.winter = NULL,
  rm.lat.equinox = FALSE, days.omit = 5, progress = TRUE, plot = TRUE,
  plot.legend = TRUE, latAllow = c(0, 10))
}
\arguments{
\item{MCMC}{a slices object.}

\item{\code{prob}}{probability estimate to return for daily positions}

\item{\code{rm.lat.equinox}}{logical remove dates from consideration around equinox?}

\item{\code{days.omit}}{integer how many days on either side to remove if \code{rm.lat.equniox == TRUE}}

\item{\code{progress}}{logical show progress bar}

\item{\code{plot}}{logical plot results upon completion}

\item{\code{plot.legend}}{logical plot legend with dates on figure}

\item{\code{latAllow}}{vector keep only consider latitudinal movements if they fall between 0 and 10 degrees in 24 hours.}
}
\value{
\code{list} DailyPositions - geographically weighted mean and associated errors
                    Schedule - arrival and departure dates of stops
                    movements \code{rasterStack} of Schedule.
}
\description{
Determine migratory schedule from archival light-level geolocator data.
}

